LayoutTests/http/tests/referrer-policy-img/no-referrer/cross-origin-http-UpgradeMixedContent.https-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/referrer-policy-img/no-referrer/cross-origin-http-UpgradeMixedContent.https.html requested insecure content from http://localhost:8000/referrer-policy/resources/image.py?expected=. This content was automatically upgraded and should be served over HTTPS.

Tests the behavior of no-referrer referrer policy when cross origin from HTTPS to HTTP, but the request is upgraded to HTTPS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is ""
PASS successfullyParsed is true

TEST COMPLETE
